🔨 refactor: patent title 재구현
