--- src/widgets/patent/styles/Title.css
@@ -1,52 +1,0 @@
-.patent__title {
-  display: flex;
-  flex-direction: column;
-
-  align-items: center;
-}
-
-.patent__title_eng {
-  font-weight: 500;
-  font-size: 0.78vmax;
-  color: var(--white);
-}
-
-.patent__title_kor {
-  font-weight: 600;
-  font-size: 2.6vmax;
-  color: #c9a96e;
-}
-
-.patent__title_description {
-  font-weight: 600;
-  font-size: 0.78vmax;
-  color: #faf7f2;
-}
-
-@media (max-width: 1024px) {
-  .patent__title_eng {
-    font-size: 1.56vmin;
-  }
-
-  .patent__title_kor {
-    font-size: 5.21vmin;
-  }
-
-  .patent__title_description {
-    font-size: 1.56vmin;
-  }
-}
-
-@media (max-width: 767px) {
-  .patent__title_eng {
-    font-size: 3.2vmin;
-  }
-
-  .patent__title_kor {
-    font-size: 8vmin;
-  }
-
-  .patent__title_description {
-    font-size: 3.2vmin;
-  }
-}
