+++ src/widgets/patent/ui/PatentTitle.tsx
@@ -1,0 +1,14 @@
+import '../styles/PatentTitle.css';
+
+export function PatentTitle() {
+  return (
+    <hgroup className='patent__title'>
+      <div className='patent__title__description'>
+        <hr />
+        <p>PATENTS</p>
+      </div>
+      <h2>특허 취득 기록</h2>
+      <h2>혁신의 증명</h2>
+    </hgroup>
+  );
+}
