--- src/widgets/patent/ui/Title.tsx
@@ -1,13 +1,0 @@
-import '../styles/Title.css';
-
-export function PatentTitle() {
-  return (
-    <hgroup className='patent__title'>
-      <p className='patent__title_eng'>PATENTS</p>
-      <h2 className='patent__title_kor'>특허 취득 기록</h2>
-      <p className='patent__title_description'>
-        끊임없는 기술 개발과 혁신의 증명
-      </p>
-    </hgroup>
-  );
-}
